Serial numbers are unsigned long, not quint32. This is the only possible reason I could imagine why it goes wrong like in bug 174866.
It would also explain why I never could reproduce this with my 32 bit CPU.
Maybe this even is the real root cause for the serial number == 0 bug??
In any case, I hope other places get the type right.

BUG: 174866

svn path=/trunk/KDE/kdepim/; revision=886618

+++ kmmsgdict.cpp
@@ -452,7 +452,7 @@ int KMMsgDict::readFolderIds( FolderStorage& storage )
   KMMsgDictREntry *rentry = new KMMsgDictREntry(count);
 
   for (unsigned int index = 0; index < count; index++) {
-    quint32 msn;
+    unsigned long msn;
 
     bool readOk = fread(&msn, sizeof(msn), 1, fp);
     if (swapByteOrder)
